tests/requirements.txt
# Testing dependencies for arpwatch-docker

# Core testing framework
pytest>=7.4.0
pytest-timeout>=2.1.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1  # For parallel test execution

# Testcontainers for integration testing
testcontainers>=4.0.0
httpx>=0.24.0

# HTTP client for API testing
requests>=2.31.0

# Docker SDK for advanced container operations
docker>=6.1.3

# Monitoring and metrics testing
prometheus-client>=0.17.1

# Network packet crafting (optional, for advanced testing)
# scapy>=2.5.0  # Uncomment if needed for network simulation

# Performance testing
locust>=2.15.1  # For load testing

# Test utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.1